Started ethernet driver and created file manager

+++ LibC/include/lemon/filesystem.h
@@ -26,7 +26,7 @@ void lemon_close(int fd);
 int lemon_read(int fd, void* buffer, size_t count);
 int lemon_write(int fd, void* buffer, size_t count);
 off_t lemon_seek(int fd, off_t offset, int whence);
-int lemon_readdir(int fd, uint32_t count, lemon_dirent_t* dirent);
+int lemon_readdir(int fd, uint64_t count, lemon_dirent_t* dirent);
 
 #ifdef __cplusplus
 }
